src/requirements.txt
fastapi
uvicorn
numpy
opencv-python
tqdm
PyYAML
requests
Pillow
google-generativeai